Move undoable reducer utility to utils directory
